feat : frontend 포인트 및 유형별 적립 포인트 구현중
