+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookCategoryController.java
@@ -4,7 +4,7 @@ import com.nhnacademy.frontend.admin.domain.request.BookCategoryCreateRequestDto
 import com.nhnacademy.frontend.admin.domain.response.BookCategoryResponseDto;
 import com.nhnacademy.frontend.admin.domain.request.BookCategoryUpdateRequestDto;
 import com.nhnacademy.frontend.admin.service.BookService;
-import com.nhnacademy.frontend.user.exception.ValidationFailedException;
+import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
