+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookTagController.java
@@ -3,7 +3,7 @@ package com.nhnacademy.frontend.admin.controller;
 import com.nhnacademy.frontend.admin.domain.request.BookTagCreateRequestDto;
 import com.nhnacademy.frontend.admin.domain.response.BookTagResponseDto;
 import com.nhnacademy.frontend.admin.service.BookService;
-import com.nhnacademy.frontend.user.exception.ValidationFailedException;
+import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
