+++ src/main/resources/templates/admin/admin-home.html
@@ -1,18 +1,35 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
 <head>
     <meta charset="UTF-8">
     <title>관리자 페이지</title>
+    <!-- Bootstrap 5 CDN -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
-<body>
-    <h1>관리자 페이지</h1>
-    <nav>
-        <ul>
-            <li><a href="/admin/books">도서 관리</a></li>
-            <li><a href="/admin/categories">카테고리 관리</a></li>
-            <li><a href="/admin/tags">태그 관리</a></li>
-            <li><a href="/">메인페이지</a></li>
+<body class="bg-light">
+<div class="container py-5">
+    <h1 class="mb-4 text-center">관리자 페이지</h1>
+    <nav class="mb-4">
+        <ul class="nav nav-pills flex-column flex-md-row justify-content-center gap-2">
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/books">도서 관리</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/categories">카테고리 관리</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/tags">태그 관리</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/pointtype">포인트 적립 규정 관리</a>
+            </li>
         </ul>
     </nav>
+    <div class="text-center">
+        <a href="/" class="btn btn-primary btn-lg">메인페이지로 이동</a>
+    </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
