+++ src/main/resources/templates/admin/book/book-list.html
@@ -1,51 +1,75 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title>관리자 도서 목록</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-<h1>도서 목록</h1>
-    <table border="1" cellpadding="8" cellspacing="0">
-        <thead>
-        <tr>
-            <th>도서 ID</th>
-            <th>제목</th>
-            <th>저자</th>
-            <th>상세 보기</th>
-            <th>관리</th>
-            <th>삭제</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="book : ${books}">
-            <td th:text="${book.id}"></td>
-            <td th:text="${book.title}"></td>
-            <td th:text="${book.author}"></td>
-            <td>
-                <a th:href="@{/admin/books/{id}(id=${book.id})}">상세</a>
-            </td>
-            <td>
-                <a th:href="@{/admin/books/{id}/tags(id=${book.id})}">태그 관리</a><br/>
-                <a th:href="@{/admin/books/{id}/categories(id=${book.id})}">카테고리 관리</a>
-            </td>
-            <td>
-                <form th:action="@{/admin/books/{id}(id=${book.id})}" method="post">
-                    <input type="hidden" name="_method" value="delete" />
-                    <button type="submit">삭제</button>
-                </form>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-
-    <a th:href="@{/admin/books/new}" class="btn">등록</a><br>
-    <a th:href="@{/admin}" class="btn">관리자 페이지</a>
-
-    <div>
-        <a th:if="${page.hasPrevious()}" th:href="@{/admin/books(page=${page.number - 1}, size=${page.size})}">이전</a>
-        <span>현재 페이지: <span th:text="${page.number + 1}">1</span></span>
-        <a th:if="${page.hasNext()}" th:href="@{/admin/books(page=${page.number + 1}, size=${page.size})}">다음</a>
+<div class="container mt-5">
+    <h1 class="mb-4">도서 목록</h1>
+    <div class="mb-3 d-flex gap-2">
+        <a th:href="@{/admin/books/new}" class="btn btn-success">도서 등록</a>
+        <a th:href="@{/admin}" class="btn btn-secondary">관리자 페이지</a>
     </div>
+    <div class="table-responsive">
+        <table class="table table-striped table-hover align-middle">
+            <thead class="table-dark">
+            <tr>
+                <th>도서 ID</th>
+                <th>제목</th>
+                <th>저자</th>
+                <th>상세 보기</th>
+                <th>관리</th>
+                <th>삭제</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="book : ${books}">
+                <td th:text="${book.id}"></td>
+                <td th:text="${book.title}"></td>
+                <td th:text="${book.author}"></td>
+                <td>
+                    <a th:href="@{/admin/books/{id}(id=${book.id})}" class="btn btn-outline-primary btn-sm">상세</a>
+                </td>
+                <td>
+                    <a th:href="@{/admin/books/{id}/tags(id=${book.id})}" class="btn btn-outline-info btn-sm mb-1">태그 관리</a><br/>
+                    <a th:href="@{/admin/books/{id}/categories(id=${book.id})}" class="btn btn-outline-warning btn-sm">카테고리 관리</a>
+                </td>
+                <td>
+                    <form th:action="@{/admin/books/{id}(id=${book.id})}" method="post" th:onsubmit="return confirm('정말 삭제하시겠습니까?');">
+                        <input type="hidden" name="_method" value="delete" />
+                        <button type="submit" class="btn btn-danger btn-sm">삭제</button>
+                    </form>
+                </td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+
+    <!-- 페이지네이션 -->
+    <nav aria-label="도서 목록 페이지네이션">
+        <ul class="pagination justify-content-center">
+            <li class="page-item" th:classappend="${!page.hasPrevious()} ? 'disabled'">
+                <a class="page-link" th:href="@{/admin/books(page=${page.number - 1}, size=${page.size})}" aria-label="이전">
+                    <span aria-hidden="true">&laquo; 이전</span>
+                </a>
+            </li>
+            <li class="page-item disabled">
+                <span class="page-link">
+                    현재 페이지: <span th:text="${page.number + 1}">1</span>
+                </span>
+            </li>
+            <li class="page-item" th:classappend="${!page.hasNext()} ? 'disabled'">
+                <a class="page-link" th:href="@{/admin/books(page=${page.number + 1}, size=${page.size})}" aria-label="다음">
+                    <span aria-hidden="true">다음 &raquo;</span>
+                </a>
+            </li>
+        </ul>
+    </nav>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
