+++ src/main/resources/templates/admin/book/category-manage.html
@@ -3,28 +3,55 @@
 <head>
     <meta charset="UTF-8" />
     <title>카테고리 관리</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-    <h1>카테고리 관리</h1>
-    <h2 th:text="'도서 ID: ' + ${book.bookId}">도서 ID</h2>
+<div class="container my-5" style="max-width: 600px;">
+    <h1 class="mb-4">카테고리 관리</h1>
+    <h5 class="mb-4" th:text="'도서 ID: ' + ${book.bookId}">도서 ID</h5>
 
-    <h2>현재 카테고리 목록</h2>
-    <ul>
-        <li th:each="category : ${book.categories}">
-            <span th:text="${category.categoryName}"></span>
-            <form th:action="@{/admin/books/{id}/categories/{categoryId}(id=${book.bookId}, categoryId=${category.categoryId})}" method="post">
-                <input type="hidden" name="_method" value="delete" />
-                <button type="submit">삭제</button>
+    <div class="card mb-4">
+        <div class="card-header bg-primary text-white">
+            현재 카테고리 목록
+        </div>
+        <ul class="list-group list-group-flush">
+            <li class="list-group-item d-flex justify-content-between align-items-center"
+                th:each="category : ${book.categories}">
+                <span th:text="${category.categoryName}"></span>
+                <form th:action="@{/admin/books/{id}/categories/{categoryId}(id=${book.bookId}, categoryId=${category.categoryId})}"
+                      method="post" class="ms-2"
+                      th:onsubmit="return confirm('이 카테고리를 삭제하시겠습니까?');">
+                    <input type="hidden" name="_method" value="delete" />
+                    <button type="submit" class="btn btn-danger btn-sm">삭제</button>
+                </form>
+            </li>
+            <li class="list-group-item text-muted" th:if="${#lists.isEmpty(book.categories)}">카테고리가 없습니다.</li>
+        </ul>
+    </div>
+
+    <div class="card mb-4">
+        <div class="card-header bg-success text-white">
+            카테고리 추가
+        </div>
+        <div class="card-body">
+            <form th:action="@{/admin/books/{id}/categories(id=${book.bookId})}" method="post" class="row g-2 align-items-center">
+                <div class="col">
+                    <input type="text" name="categoryId" class="form-control" placeholder="카테고리 아이디 입력" required />
+                </div>
+                <div class="col-auto">
+                    <button type="submit" class="btn btn-success">추가</button>
+                </div>
             </form>
-        </li>
-    </ul>
+        </div>
+    </div>
 
-    <h3>카테고리 추가</h3>
-    <form th:action="@{/admin/books/{id}/categories(id=${book.bookId})}" method="post">
-        <input type="text" name="categoryId" placeholder="카테고리 아이디 입력" required />
-        <button type="submit">추가</button>
-    </form>
-    <a th:href="@{/admin/books/{bookId}(bookId=${book.bookId})}">이전 페이지</a><br>
-    <a href="/admin/books">도서 목록으로</a>
+    <div class="d-flex gap-2">
+        <a th:href="@{/admin/books/{bookId}(bookId=${book.bookId})}" class="btn btn-secondary">이전 페이지</a>
+        <a href="/admin/books" class="btn btn-outline-primary">도서 목록으로</a>
+    </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
