+++ src/main/resources/templates/admin/book/detail.html
@@ -1,108 +1,117 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
   <meta charset="UTF-8">
   <title th:text="'책 상세 - ' + ${book.title}">책 상세</title>
+  <!-- Bootstrap CSS -->
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-<h1 th:text="'책 상세 정보: ' + ${book.title}">책 상세 정보</h1>
+<div class="container my-5" style="max-width: 900px;">
+  <h1 class="mb-4" th:text="'책 상세 정보: ' + ${book.title}">책 상세 정보</h1>
 
-<table border="1" cellpadding="8" cellspacing="0">
-  <tbody>
-  <tr>
-    <th>아이디</th>
-    <td th:text="${book.id}"></td>
-  </tr>
-  <tr>
-    <th>책 표지</th>
-    <td>
-      <img th:src="${book.image}" alt="책 표지" style="max-width: 200px; max-height: 300px;" />
-    </td>
-  </tr>
-  <tr>
-    <th>제목</th>
-    <td th:text="${book.title}"></td>
-  </tr>
-  <tr>
-    <th>목차</th>
-    <td th:text="${book.toc}"></td>
-  </tr>
-  <tr>
-    <th>설명</th>
-    <td th:text="${book.description}"></td>
-  </tr>
-  <tr>
-    <th>저자</th>
-    <td th:text="${book.author}"></td>
-  </tr>
-  <tr>
-    <th>출판사</th>
-    <td th:text="${book.publisher}"></td>
-  </tr>
-  <tr>
-    <th>출판일</th>
-    <td th:text="${book.publishAt}"></td>
-  </tr>
-  <tr>
-    <th>ISBN</th>
-    <td th:text="${book.isbn}"></td>
-  </tr>
-  <tr>
-    <th>원가</th>
-    <td th:text="${book.originalPrice} + '원'"></td>
-  </tr>
-  <tr>
-    <th>판매가</th>
-    <td th:text="${book.salePrice} + '원'"></td>
-  </tr>
-  <tr>
-    <th>포장 여부</th>
-    <td th:text="${book.wrappable ? '가능' : '불가능'}"></td>
-  </tr>
-  <tr>
-    <th>재고</th>
-    <td th:text="${book.stock} + ' 권'"></td>
-  </tr>
-  <tr>
-    <th>책 상태</th>
-    <td th:text="${book.status}"></td>
-  </tr>
-  <tr>
-    <th>등록일</th>
-    <td th:text="${book.createAt}"></td>
-  </tr>
-  <tr>
-    <th>수정일</th>
-    <td th:text="${book.updateAt}"></td>
-  </tr>
-  <tr>
-    <th>카테고리</th>
-    <td>
-      <ul>
-        <li th:each="category : ${book.bookCategories}" th:text="${category.categoryName}" style="list-style-type:none;"></li>
-      </ul>
-    </td>
-  </tr>
-  <tr>
-    <th>태그</th>
-    <td>
-      <ul>
-        <li th:each="tag : ${book.bookTags}" th:text="${tag.tagName}" style="list-style-type:none"></li>
-      </ul>
-    </td>
-  </tr>
-  </tbody>
-</table>
+  <table class="table table-bordered align-middle">
+    <tbody>
+    <tr>
+      <th class="table-secondary" style="width: 140px;">아이디</th>
+      <td th:text="${book.id}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">책 표지</th>
+      <td>
+        <img th:src="${book.image}" alt="책 표지" class="img-thumbnail" style="max-width: 200px; max-height: 300px;" />
+      </td>
+    </tr>
+    <tr>
+      <th class="table-secondary">제목</th>
+      <td th:text="${book.title}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">목차</th>
+      <td th:text="${book.toc}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">설명</th>
+      <td th:text="${book.description}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">저자</th>
+      <td th:text="${book.author}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">출판사</th>
+      <td th:text="${book.publisher}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">출판일</th>
+      <td th:text="${book.publishAt}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">ISBN</th>
+      <td th:text="${book.isbn}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">원가</th>
+      <td th:text="${book.originalPrice} + '원'"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">판매가</th>
+      <td th:text="${book.salePrice} + '원'"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">포장 여부</th>
+      <td th:text="${book.wrappable ? '가능' : '불가능'}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">재고</th>
+      <td th:text="${book.stock} + ' 권'"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">책 상태</th>
+      <td th:text="${book.status}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">등록일</th>
+      <td th:text="${book.createAt}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">수정일</th>
+      <td th:text="${book.updateAt}"></td>
+    </tr>
+    <tr>
+      <th class="table-secondary">카테고리</th>
+      <td>
+        <ul class="list-inline mb-0">
+          <li class="list-inline-item badge bg-primary"
+              th:each="category : ${book.bookCategories}"
+              th:text="${category.categoryName}"></li>
+        </ul>
+      </td>
+    </tr>
+    <tr>
+      <th class="table-secondary">태그</th>
+      <td>
+        <ul class="list-inline mb-0">
+          <li class="list-inline-item badge bg-info text-dark"
+              th:each="tag : ${book.bookTags}"
+              th:text="${tag.tagName}"></li>
+        </ul>
+      </td>
+    </tr>
+    </tbody>
+  </table>
 
-<a th:href="@{/admin/books/{id}/edit(id=${book.id})}">
-  <button type="button">수정</button>
-</a>
-
-<form th:action="@{/admin/books/{id}(id=${book.id})}" method="post">
-  <input type="hidden" name="_method" value="delete" />
-  <button type="submit">삭제</button>
-</form>
-
-<a href="/admin/books">책 목록으로</a>
+  <div class="d-flex gap-2 my-4">
+    <a th:href="@{/admin/books/{id}/edit(id=${book.id})}" class="btn btn-warning">수정</a>
+    <form th:action="@{/admin/books/{id}(id=${book.id})}" method="post"
+          th:onsubmit="return confirm('정말 삭제하시겠습니까?');" class="d-inline">
+      <input type="hidden" name="_method" value="delete" />
+      <button type="submit" class="btn btn-danger">삭제</button>
+    </form>
+    <a href="/admin/books" class="btn btn-secondary">책 목록으로</a>
+  </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
