+++ src/main/resources/templates/admin/book/tag-manage.html
@@ -1,30 +1,57 @@
 <!DOCTYPE html>
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
-  <meta charset="UTF-8" />
-  <title>태그 관리</title>
+    <meta charset="UTF-8" />
+    <title>태그 관리</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-    <h1>태그 관리</h1>
-    <h2 th:text="'도서 ID: ' + ${book.bookId}">도서 ID</h2>
+<div class="container my-5" style="max-width: 600px;">
+    <h1 class="mb-4">태그 관리</h1>
+    <h5 class="mb-4" th:text="'도서 ID: ' + ${book.bookId}">도서 ID</h5>
 
-    <h2>현재 태그 목록</h2>
-    <ul id="tag-list">
-      <li th:each="tag : ${book.tags}">
-        <span th:text="${tag.tagName}"></span>
-        <form th:action="@{/admin/books/{id}/tags/{tagId}(id=${book.bookId}, tagId=${tag.tagId})}" method="post">
-          <input type="hidden" name="_method" value="delete" />
-          <button type="submit">삭제</button>
-        </form>
-      </li>
-    </ul>
+    <div class="card mb-4">
+        <div class="card-header bg-info text-dark">
+            현재 태그 목록
+        </div>
+        <ul class="list-group list-group-flush">
+            <li class="list-group-item d-flex justify-content-between align-items-center"
+                th:each="tag : ${book.tags}">
+                <span th:text="${tag.tagName}"></span>
+                <form th:action="@{/admin/books/{id}/tags/{tagId}(id=${book.bookId}, tagId=${tag.tagId})}"
+                      method="post" class="ms-2"
+                      th:onsubmit="return confirm('이 태그를 삭제하시겠습니까?');">
+                    <input type="hidden" name="_method" value="delete" />
+                    <button type="submit" class="btn btn-danger btn-sm">삭제</button>
+                </form>
+            </li>
+            <li class="list-group-item text-muted" th:if="${#lists.isEmpty(book.tags)}">태그가 없습니다.</li>
+        </ul>
+    </div>
 
-    <h3>태그 추가</h3>
-    <form th:action="@{/admin/books/{id}/tags(id=${book.bookId})}" method="post">
-      <input type="text" name="tagId" placeholder="태그 ID 입력" required />
-      <button type="submit">추가</button>
-    </form>
-    <a th:href="@{/admin/books/{bookId}(bookId=${book.bookId})}">이전 페이지</a><br>
-    <a href="/admin/books">도서 목록으로</a>
+    <div class="card mb-4">
+        <div class="card-header bg-success text-white">
+            태그 추가
+        </div>
+        <div class="card-body">
+            <form th:action="@{/admin/books/{id}/tags(id=${book.bookId})}" method="post" class="row g-2 align-items-center">
+                <div class="col">
+                    <input type="text" name="tagId" class="form-control" placeholder="태그 ID 입력" required />
+                </div>
+                <div class="col-auto">
+                    <button type="submit" class="btn btn-success">추가</button>
+                </div>
+            </form>
+        </div>
+    </div>
+
+    <div class="d-flex gap-2">
+        <a th:href="@{/admin/books/{bookId}(bookId=${book.bookId})}" class="btn btn-secondary">이전 페이지</a>
+        <a href="/admin/books" class="btn btn-outline-primary">도서 목록으로</a>
+    </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
