+++ src/main/resources/templates/admin/book/update-form.html
@@ -1,64 +1,78 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title th:text="'도서 수정 - ' + ${book.title}">도서 수정</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-<h1 th:text="'도서 수정: ' + ${book.title}">도서 수정</h1>
+<div class="container my-5" style="max-width: 700px;">
+    <h1 class="mb-4" th:text="'도서 수정: ' + ${book.title}">도서 수정</h1>
+    <form th:action="@{/admin/books/{id}(id=${bookId})}" method="post">
+        <input type="hidden" name="_method" value="put" />
 
-<form th:action="@{/admin/books/{id}(id=${bookId})}" method="post">
-    <input type="hidden" name="_method" value="put" />
-
-    <label>제목</label><br/>
-    <input type="text" name="title" th:value="${book.title}" required/><br/><br/>
-
-    <labe>이미지</labe><br/>
-    <input type="text" name="image" th:value="${book.image}" required/><br/><br/>
-
-    <label>설명</label><br/>
-    <textarea name="description" rows="4" cols="50" th:text="${book.description}" required></textarea><br/><br/>
-
-    <label>목차</label><br/>
-    <textarea name="toc" rows="4" cols="50" th:text="${book.toc}"></textarea><br/><br/>
-
-    <label>저자</label><br/>
-    <input type="text" name="author" th:value="${book.author}" required/><br/><br/>
-
-    <label>출판사</label><br/>
-    <input type="text" name="publisher" th:value="${book.publisher}" required/><br/><br/>
-
-    <label>출판일</label><br/>
-    <input type="date" name="publishAt"
-           th:value="${book.publishAt}"/><br/><br/>
-
-
-    <label>원가</label><br/>
-    <input type="number" name="originalPrice" th:value="${book.originalPrice}" min="0" required/><br/><br/>
-
-    <label>판매가</label><br/>
-    <input type="number" name="salePrice" th:value="${book.salePrice}" min="0" required/><br/><br/>
-
-    <label>포장 가능 여부</label><br/>
-    <select name="wrappable">
-        <option th:value="true" th:selected="${book.wrappable}">가능</option>
-        <option th:value="false" th:selected="${!book.wrappable}">불가능</option>
-    </select><br/><br/>
-
-    <label>재고</label><br/>
-    <input type="number" name="stock" th:value="${book.stock}" min="0"/><br/><br/>
-
-    <label>책 상태</label><br/>
-    <select name="status">
-        <option th:value="ON_SALE" th:selected="${book.status == 'ON_SALE'}">판매중</option>
-        <option th:value="SALE_END" th:selected="${book.status == 'SALE_END'}">판매종료</option>
-    </select><br/><br/>
-
-    <button type="submit">수정 완료</button>
-</form>
-
-<br/>
-<a href="/admin/books" style="text-decoration:none;">목록으로</a>
-
+        <div class="mb-3">
+            <label class="form-label">제목</label>
+            <input type="text" name="title" th:value="${book.title}" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">이미지</label>
+            <input type="text" name="image" th:value="${book.image}" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">설명</label>
+            <textarea name="description" rows="4" class="form-control" th:text="${book.description}" required></textarea>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">목차</label>
+            <textarea name="toc" rows="4" class="form-control" th:text="${book.toc}"></textarea>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">저자</label>
+            <input type="text" name="author" th:value="${book.author}" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">출판사</label>
+            <input type="text" name="publisher" th:value="${book.publisher}" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">출판일</label>
+            <input type="date" name="publishAt" th:value="${book.publishAt}" class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">원가</label>
+            <input type="number" name="originalPrice" th:value="${book.originalPrice}" min="0" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">판매가</label>
+            <input type="number" name="salePrice" th:value="${book.salePrice}" min="0" required class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">포장 가능 여부</label>
+            <select name="wrappable" class="form-select">
+                <option th:value="true" th:selected="${book.wrappable}">가능</option>
+                <option th:value="false" th:selected="${!book.wrappable}">불가능</option>
+            </select>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">재고</label>
+            <input type="number" name="stock" th:value="${book.stock}" min="0" class="form-control"/>
+        </div>
+        <div class="mb-3">
+            <label class="form-label">책 상태</label>
+            <select name="status" class="form-select">
+                <option th:value="ON_SALE" th:selected="${book.status == 'ON_SALE'}">판매중</option>
+                <option th:value="SALE_END" th:selected="${book.status == 'SALE_END'}">판매종료</option>
+            </select>
+        </div>
+        <div class="d-flex gap-2">
+            <button type="submit" class="btn btn-primary">수정 완료</button>
+            <a href="/admin/books" class="btn btn-secondary">목록으로</a>
+        </div>
+    </form>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
