+++ src/main/resources/templates/admin/category/create-form.html
@@ -1,27 +1,38 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
-    <meta charset="UTF-8">
-    <title>카테고리 생성</title>
+  <meta charset="UTF-8">
+  <title>카테고리 생성</title>
+  <!-- Bootstrap 5 CDN -->
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
-<body>
-  <h1>카테고리 생성</h1>
-  <form th:action="@{/admin/categories}" th:object="${request}" method="post">
-    <div>
-      <label for="categoryName">카테고리 이름:</label>
-      <input type="text" th:field="*{categoryName}" id="categoryName" required />
+<body class="bg-light">
+<div class="container my-5">
+  <div class="row justify-content-center">
+    <div class="col-md-6 col-lg-5">
+      <div class="card shadow-sm">
+        <div class="card-body">
+          <h1 class="card-title text-center mb-4">카테고리 생성</h1>
+          <form th:action="@{/admin/categories}" th:object="${request}" method="post">
+            <div class="mb-3">
+              <label for="categoryName" class="form-label">카테고리 이름</label>
+              <input type="text" class="form-control" th:field="*{categoryName}" id="categoryName" required />
+            </div>
+            <div class="mb-4">
+              <label for="parentId" class="form-label">부모 카테고리 ID</label>
+              <input type="number" class="form-control" th:field="*{parentId}" id="parentId" />
+            </div>
+            <button type="submit" class="btn btn-primary w-100">생성</button>
+          </form>
+          <div class="mt-3 text-center">
+            <a href="/admin/categories" class="btn btn-secondary">카테고리 목록</a>
+          </div>
+        </div>
+      </div>
     </div>
-
-    <div>
-      <label for="parentId">부모 카테고리 ID :</label>
-      <input type="number" th:field="*{parentId}" id="parentId" />
-    </div>
-
-    <div>
-      <button type="submit">생성</button>
-    </div>
-  </form>
-
-  <a href="/admin/categories">카테고리 목록</a>
+  </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
