--- src/main/resources/templates/admin/main.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>관리자페이지</title>
-</head>
-<body>
-<h1>관리자페이지</h1>
-
-<ul>
-    <li>
-        <a href="/admin/pointtype">포인트 적립 규정 관리</a>
-    </li>
-</ul>
