+++ src/main/resources/templates/admin/pointtype/pointType-edit.html
