+++ src/main/resources/templates/admin/pointtype/pointType-register.html
