+++ src/main/resources/templates/admin/tag/create-form.html
@@ -1,18 +1,35 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title>태그 생성</title>
+    <!-- Bootstrap 5 CDN -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
-<body>
-    <h1>도서 태그 생성</h1>
-    <form th:action="@{/admin/tags}" method="post" th:object="${request}">
-
-        <label for="tagName">태그 이름:</label>
-        <input type="text" id="tagName" th:field="*{tagName}" placeholder="태그 이름 입력" required />
-
-        <button type="submit">생성</button>
-    </form>
-    <a href="/admin/tags">태그 목록</a>
+<body class="bg-light">
+<div class="container my-5">
+    <div class="row justify-content-center">
+        <div class="col-md-6 col-lg-5">
+            <div class="card shadow-sm">
+                <div class="card-body">
+                    <h1 class="card-title text-center mb-4">도서 태그 생성</h1>
+                    <form th:action="@{/admin/tags}" method="post" th:object="${request}">
+                        <div class="mb-3">
+                            <label for="tagName" class="form-label">태그 이름</label>
+                            <input type="text" class="form-control" id="tagName"
+                                   th:field="*{tagName}" placeholder="태그 이름 입력" required>
+                        </div>
+                        <button type="submit" class="btn btn-primary w-100">생성</button>
+                    </form>
+                    <div class="mt-3 text-center">
+                        <a href="/admin/tags" class="btn btn-secondary">태그 목록</a>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
