+++ src/main/resources/templates/book/book-detail.html
@@ -3,70 +3,78 @@
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       layout:decorate="~{layout/layout}">
 <head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
   <title layout:fragment="title" th:text="${book.title} + ' 상세정보'">도서 상세</title>
+  <!-- Bootstrap CSS -->
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
 <div layout:fragment="content" class="container mt-4" style="max-width:800px;">
-  <div class="card">
+  <div class="card shadow-sm">
     <img th:src="${book.image}" class="card-img-top" alt="Book Image" style="height:250px; object-fit:cover;">
     <div class="card-body">
-      <h3 class="card-title" th:text="${book.title}">도서 제목</h3>
-      <div>
+      <h3 class="card-title mb-3" th:text="${book.title}">도서 제목</h3>
+      <div class="mb-3">
         <h5>카테고리</h5>
-        <ul>
-          <li th:each="cat : ${book.bookCategories}" th:text="${cat.categoryName}">카테고리명</li>
+        <ul class="list-group list-group-flush">
+          <li class="list-group-item" th:each="cat : ${book.bookCategories}" th:text="${cat.categoryName}">카테고리명</li>
         </ul>
       </div>
-      <div>
+      <div class="mb-3">
         <h5>태그</h5>
-        <ul>
-          <li th:each="tag : ${book.bookTags}" th:text="${tag.tagName}">태그명</li>
+        <ul class="list-inline">
+          <li class="list-inline-item badge bg-secondary" th:each="tag : ${book.bookTags}" th:text="${tag.tagName}">태그명</li>
         </ul>
       </div>
-      <p>저자: <span th:text="${book.author}">저자명</span></p>
-      <p>출판사: <span th:text="${book.publisher}">출판사명</span></p>
-      <p>출판일: <span th:text="${book.publishAt}"></span></p>
-      <p>정가: <span th:text="${book.originalPrice}">0</span>원</p>
-      <p>판매가: <span th:text="${book.salePrice}">0</span>원</p>
-      <p>도서 상태: <span th:text="${book.status}"></span></p>
-      <p>재고: <span th:text="${book.stock}">0</span>권</p>
-      <p>포장 가능 여부: <span th:text="${book.wrappable} ? '가능' : '불가'"></span></p>
-
-      <p th:if="${book.toc}">
-        목차: <span th:text="${book.toc}"></span>
-      </p>
-      <p th:if="${book.description}">
-        설명: <span th:text="${book.description}"></span>
-      </p>
-
-      <p>좋아요 수: <span th:text="${book.likeCount}"></span>개</p>
-      <!-- 좋아요 버튼 -->
-      <form th:action="@{'/books/' + ${book.id} + '/bookLikes'}" method="post" style="display:inline;">
-        <input type="hidden" name="_method" value="post" />
-        <button type="submit" class="btn btn-primary">좋아요</button>
-      </form>
-
-      <form th:action="@{'/books/' + ${book.id} + '/bookLikes'}" method="post" style="display:inline;"
-            onsubmit="return confirm('좋아요를 취소하시겠습니까?');">
-        <input type="hidden" name="_method" value="delete" />
-        <button type="submit" class="btn btn-secondary">좋아요 취소</button>
-      </form>
-
-      <form th:action="@{/books}" method="post" class="mt-3">
+      <ul class="list-group list-group-flush mb-3">
+        <li class="list-group-item">저자: <span th:text="${book.author}">저자명</span></li>
+        <li class="list-group-item">출판사: <span th:text="${book.publisher}">출판사명</span></li>
+        <li class="list-group-item">출판일: <span th:text="${book.publishAt}"></span></li>
+        <li class="list-group-item">정가: <span th:text="${book.originalPrice}">0</span>원</li>
+        <li class="list-group-item">판매가: <span th:text="${book.salePrice}">0</span>원</li>
+        <li class="list-group-item">도서 상태: <span th:text="${book.status}"></span></li>
+        <li class="list-group-item">재고: <span th:text="${book.stock}">0</span>권</li>
+        <li class="list-group-item">포장 가능 여부: <span th:text="${book.wrappable} ? '가능' : '불가'"></span></li>
+      </ul>
+      <div th:if="${book.toc}" class="mb-2">
+        <strong>목차:</strong> <span th:text="${book.toc}"></span>
+      </div>
+      <div th:if="${book.description}" class="mb-2">
+        <strong>설명:</strong> <span th:text="${book.description}"></span>
+      </div>
+      <div class="mb-3">
+        <span class="me-2">좋아요 수: <span th:text="${book.likeCount}"></span>개</span>
+        <!-- 좋아요 버튼 -->
+        <form th:action="@{'/books/' + ${book.id} + '/bookLikes'}" method="post" class="d-inline">
+          <input type="hidden" name="_method" value="post" />
+          <button type="submit" class="btn btn-primary btn-sm">좋아요</button>
+        </form>
+        <form th:action="@{'/books/' + ${book.id} + '/bookLikes'}" method="post" class="d-inline"
+              onsubmit="return confirm('좋아요를 취소하시겠습니까?');">
+          <input type="hidden" name="_method" value="delete" />
+          <button type="submit" class="btn btn-outline-secondary btn-sm">좋아요 취소</button>
+        </form>
+      </div>
+      <form th:action="@{/books}" method="post" class="mt-3 row g-3 align-items-center">
         <input type="hidden" name="bookId" th:value="${book.id}" />
         <input type="hidden" name="title" th:value="${book.title}" />
         <input type="hidden" name="salePrice" th:value="${book.salePrice}" />
         <input type="hidden" name="wrappable" th:value="${book.wrappable}" />
-
-        <div class="mb-2">
-          <label for="quantity">수량</label>
+        <div class="col-auto">
+          <label for="quantity" class="col-form-label">수량</label>
+        </div>
+        <div class="col-auto">
           <input type="number" id="quantity" name="quantity" value="1" min="1" max="100" class="form-control" style="width:120px;" required />
         </div>
-
-        <button type="submit" class="btn btn-success">구매하기</button>
+        <div class="col-auto">
+          <button type="submit" class="btn btn-success">구매하기</button>
+        </div>
       </form>
     </div>
   </div>
 </div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
