+++ src/main/resources/templates/home.html
@@ -1,13 +1,15 @@
-<!-- main.html -->
 <!DOCTYPE html>
 <html lang="ko" xmlns:th="http://www.thymeleaf.org"
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       layout:decorate="~{layout/layout}">
 <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
     <title layout:fragment="title">메인 페이지</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
-<!-- ✅ 본문 콘텐츠 fragment -->
 <div layout:fragment="content">
     <div class="container mt-4" style="max-width:1200px;">
         <div class="row">
@@ -25,24 +27,28 @@
             </div>
             <!-- 세로 메뉴 -->
             <div class="col-2 d-flex flex-column align-items-center">
-                <div class="bg-secondary text-white d-flex align-items-center justify-content-center" style="height:260px; writing-mode:vertical-rl; text-orientation: mixed; font-size:1.1rem;">
+                <div class="bg-secondary text-white d-flex align-items-center justify-content-center"
+                     style="height:260px; writing-mode:vertical-rl; text-orientation: mixed; font-size:1.1rem;">
                     <span>분류<br>테마<br>추천<br>이벤트</span>
                 </div>
             </div>
             <!-- 책 2열 -->
-            <div class="row row-cols-2 g-4">
-                <div class="col" th:each="book : ${books}">
-                    <div class="card h-100">
-                        <a th:href="@{/books/{id}(id=${book.id})}">
-                            <img th:src="${book.image}" class="card-img-top" alt="Book Image" style="height:180px; object-fit:cover;">
-                        </a>
-                        <div class="card-body">
-                            <h5 class="card-title" th:text="${book.title}">도서 제목</h5>
-                            <p class="card-text">
-                                저자: <span th:text="${book.author}">저자</span><br>
-                                가격: <span th:text="${book.salePrice}">0</span>원<br>
-                                재고: <span th:text="${book.stock}">0</span>권
-                            </p>
+            <div class="col-8">
+                <div class="row row-cols-2 g-4">
+                    <div class="col" th:each="book : ${books}">
+                        <div class="card h-100">
+                            <a th:href="@{/books/{id}(id=${book.id})}">
+                                <img th:src="${book.image}" class="card-img-top" alt="Book Image"
+                                     style="height:180px; object-fit:cover;">
+                            </a>
+                            <div class="card-body">
+                                <h5 class="card-title" th:text="${book.title}">도서 제목</h5>
+                                <p class="card-text">
+                                    저자: <span th:text="${book.author}">저자</span><br>
+                                    가격: <span th:text="${book.salePrice}">0</span>원<br>
+                                    재고: <span th:text="${book.stock}">0</span>권
+                                </p>
+                            </div>
                         </div>
                     </div>
                 </div>
@@ -56,14 +62,12 @@
                     <li class="page-item" th:classappend="${page.number == 0} ? 'disabled'">
                         <a class="page-link" th:href="@{/(page=${page.number - 1}, size=${page.size})}" aria-label="Previous">&larr;</a>
                     </li>
-
                     <!-- 페이지 번호 -->
                     <li class="page-item"
                         th:each="i : ${#numbers.sequence(0, page.totalPages - 1)}"
                         th:classappend="${i == page.number} ? 'active'">
                         <a class="page-link" th:href="@{/(page=${i}, size=${page.size})}" th:text="${i + 1}">1</a>
                     </li>
-
                     <!-- 다음 버튼 -->
                     <li class="page-item" th:classappend="${page.number == page.totalPages - 1} ? 'disabled'">
                         <a class="page-link" th:href="@{/(page=${page.number + 1}, size=${page.size})}" aria-label="Next">&rarr;</a>
@@ -73,5 +77,7 @@
         </div>
     </div>
 </div>
+<!-- Bootstrap JS (optional) -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
