+++ src/main/resources/templates/layout/layout.html
@@ -26,7 +26,7 @@
             <span class="me-3 text-white" th:text="${loginUserId + ' 님 환영합니다!'}"></span>
 
             <!-- 관리자만 보이는 버튼 -->
-            <a href="/admin/main" sec:authorize="hasRole('ROLE_ADMIN')" class="btn btn-warning btn-sm me-3">관리자페이지</a>
+            <a href="/admin" sec:authorize="hasRole('ROLE_ADMIN')" class="btn btn-warning btn-sm me-3">관리자페이지</a>
 
             <!-- 오른쪽: 로그아웃 버튼 (inline-block 설정) -->
             <form th:action="@{/auth/logout}" method="post" class="mb-0 d-inline">
